requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
python-dotenv
mercadopago
uuid
cryptography